Add an alt attribute to artboard image
So I'm proposing this because I did a light accessibility audit of NBC News' graphics and this is one of the problems that came up -- ai2html images don't have alt attributes. An empty alt attribute is better than no alt attribute -- without an alt attribute, screen readers read the src attribute aloud to the reader.

Thank you for considering this change.